updated to SDK v6.1.1
