+++ Xamarin.Smatto.Test/Resources/layout/Main.axml
@@ -2,17 +2,15 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="vertical"
     android:layout_width="fill_parent"
-    android:layout_height="fill_parent"
-    >
-<Button  
-    android:id="@+id/MyButton"
-    android:layout_width="fill_parent" 
-    android:layout_height="wrap_content" 
-    android:text="@string/Hello"
-    />
-  <LinearLayout
-      android:id="@+id/ad_container"
-      android:layout_width="fill_parent"
-      android:layout_height="wrap_content"
-      android:gravity="center_horizontal" />
+    android:layout_height="fill_parent">
+  <TextView
+    android:text="Banner Ad"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+     />
+    <RelativeLayout
+        android:id="@+id/ad_container"
+        android:layout_width="fill_parent"
+        android:layout_height="250dp"
+        />    
 </LinearLayout>
