--- Xamarin.Smatto/Properties/AssemblyInfo.cs
@@ -1,15 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle("Xamarin.Smatto")]
-[assembly: AssemblyDescription("")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("Studio Nine")]
-[assembly: AssemblyProduct("Xamarin.Smatto")]
-[assembly: AssemblyCopyright("Copyright © Studio Nine 2015")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-[assembly: ComVisible(false)]
-
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
