--- Xamarin.Smatto/Transforms/Metadata.xml
@@ -1,10 +1,0 @@
-<metadata>
-  <remove-node path="/api/package[@name='com.smaato.soma.video']" />
-  <remove-node path="/api/package[@name='com.smaato.soma.internal']" />
-  <remove-node path="/api/package[@name='com.smaato.soma.internal.requests']" />
-  <remove-node path="/api/package[@name='com.smaato.soma.internal.requests.reports']" />
-  <remove-node path="/api/package[@name='com.smaato.soma.internal.utilities']" />
-
-  <remove-node path="/api/package[@name='com.smaato.soma']/interface[@name='BaseViewInterface']" />
-
-</metadata>
